chore: Renamed core/server/session.service.mts to core/server/session.mts.
